service/requirements.txt
munch
redis
opencv-python
numpy
Pillow
flask
flask_restx
flask_healthz
werkzeug